+++ test/cookbooks/test/recipes/server_install.rb
@@ -1,7 +1,0 @@
-hashicorp_vault_install 'package' do
-  ui true
-  disable_performance_standby true
-  tls_disable true
-  sensitive true
-  action [:install]
-end