+++ test/fixtures/policies/default.rb
@@ -1,5 +1,0 @@
-name 'default'
-default_source :supermarket
-default_source :chef_repo, '..'
-cookbook 'hashicorp-vault', path: '../../..'
-run_list 'hashicorp-vault::default'